+++ lib/features/screen_home/presentation/lineup/lineup.dart
@@ -3,11 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:soundboard/features/screen_home/presentation/lineup/classes/class_lineup.dart';
 
 class LineupSection extends StatelessWidget {
-  final double availableWidth;
-
   const LineupSection({
     super.key,
-    required this.availableWidth,
   });
 
   @override
@@ -17,6 +14,7 @@ class LineupSection extends StatelessWidget {
         builder: (context, constraints) {
           return Lineup(
             availableWidth: constraints.maxWidth,
+            availableHeight: constraints.maxHeight,
           );
         },
       ),